first section algorithm
