+++ src/index.html
@@ -11,39 +11,53 @@
     <!-- Personalized Scripts -->
     <script>
         const initOptI18N = {
-            lng: "it",
+            /* lng: "it", */
             fallbackLng: 'it',
             debug: true,
             whiteList: ['it', 'en'],
+            ns: "translation",
             backend: {
                 // load from i18next-gitbook repo
                 loadPath: "./locales/{{lng}}/{{ns}}.json",
                 crossDomain: true,
-            }
+            },
         };
     </script>
 </head>
 
 <body>
-    <div style="height: 150px">
+    <div style="height: 150px;">
         <button onClick="I18N.changeLng('en')">
             english
         </button>
         <button onClick="I18N.changeLng('it')">
             italian
         </button>
-        <hr />
+        <!--         <hr />
         <input data-lang data-lang-alt="T_saluto" data-lang-placeholder="T_info"></input>
         <input data-lang data-lang-alt-placeholder="T_saluto"></input>
         <div data-lang="T_info"></div>
-        <div data-lang-it="T_info"></div>
+
+        <hr /> -->
+
+        <div data-lang="T_saluto;1" style="border:1px solid green">
+            <div style="border:1px solid green">Testo da non tradurre 1</div>
+            TESTO DA TRADURRE
+            <div style="border:1px solid green">Testo da non tradurre 2</div>
+        </div>
 
         <hr />
-        <!--   <div data-lang="T_info"></div> -->
 
-        <div id="automationTranslation">
+        <!--        <div data-lang="T_info" style="border:1px solid red" id="automationTranslation">
 
-        </div>
+            <div style="border:1px solid blue" data-lang="T_saluto">T_saluto</div>
+            Testo da sostituire con la traduzione
+            <div style="border:1px solid blue" data-lang="T_saluto">T_saluto</div>
+
+        </div> -->
+
+
+
     </div>
 
 </body>
